screener/requirements.txt
npyscreen==4.10.5
requests==2.32.3
python-dotenv==1.1.0